Adjust exports & naming.
