+++ packages/ui/react-ui-stack/src/components/index.ts
@@ -4,4 +4,3 @@
 
 export * from './Stack';
 export * from './StackItem';
-export * from './PlankHeading';
